fix: error when delete multiple files at the same time

+++ packages/nextjs/src/index.ts
@@ -500,8 +500,9 @@ export default (ctx: IPluginContext, pluginOpts: PluginOptions) => {
                         }
 
                         if (operation === 'removed') {
-                            fs.rmSync(outputFilePath)
-
+                            if (fs.existsSync(outputFilePath)) {
+                                fs.rmSync(outputFilePath)
+                            }
                             const dir = path.dirname(filePath)
                             const ext = path.extname(filePath)
                             const base = path.basename(filePath, ext)
